apps/report-manager/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
pydantic==2.11.9
python-multipart==0.0.9
psycopg2-binary==2.9.9
sqlalchemy==2.0.45
python-dotenv==1.2.1
minio==7.2.7
PyPDF2==3.0.1
Pillow==12.0.0
pytesseract==0.3.10
requests==2.32.5
werkzeug==3.0.6
email-validator==2.2.0
pdf2image==1.17.0
json-repair==0.33.0
paddlepaddle==2.6.2
paddleocr==2.8.1
